Task1.2: Move random generator function to the separate file
