+++ task1-simple-program/guess.h
@@ -1,0 +1,6 @@
+#ifndef GUESS_H
+#define GUESS_H
+
+int get_rand(void);
+
+#endif //GUESS_H
